Continued work on scene reordering, half finished edit/delete of scenes.

+++ social.py
@@ -20,20 +20,19 @@ social = Blueprint('social', __name__, template_folder='templates')
 @auth.login_required
 def index():
     user = auth.get_logged_in_user()
-    following_messages = Message.select().where( Q(conversation=0) & Q(user__in=user.following()) 
-    ).order_by(('pub_date', 'desc'))
+    following_messages = Message.select().where(Message.conversation ==0, Message.user << user.following()).order_by(Message.pub_date.desc())
     return object_list('social/social.html', following_messages, 'following_message_list')
 
 @social.route('/public/')
 def public_timeline():
-    messages = Message.select().where(conversation=0).order_by(('pub_date', 'desc'))
+    messages = Message.select().where(Message.conversation == 0).order_by(Message.pub_date.desc())
     return object_list('social/public_messages.html', messages, 'message_list')
 
 @social.route('/conversations/', methods=['GET'])
 @auth.login_required
 def conversations():
     user = auth.get_logged_in_user()
-    conversations = Conversation.select().join(ConversationMembers).where( member=user)
+    conversations = Conversation.select().join(ConversationMembers).where( ConversationMembers.member == user)
     return object_list('social/conversations.html', conversations, 'conversation_list')
 
 @social.route('/conversations/new/', methods=['GET', 'POST'])
@@ -49,7 +48,7 @@ def conversation_new():
         recipients = []
         for r in request.args.getlist('recipients'):
             recipients.extend([r.strip() for r in r.split(',')])
-        r_query = User.select().where(username__in=recipients)
+        r_query = User.select().where(User.username << recipients) # In recipients
         recipients = [r for r in r_query] # Iterate now to get a list, not a query object
         print "Found recipients %s from request param %s" % (recipients, request.args.getlist('recipients'))
         # Get most recent conversation if there is one
@@ -69,7 +68,7 @@ def conversation_new():
         # request.form is a MultiDict, which allows one or many values per key
         # we want to get all values for key recipients, so we use getlist()
         recipients = request.form.getlist('recipients')
-        r_query = User.select().where(username__in=request.form.getlist('recipients'))
+        r_query = User.select().where(User.username << request.form.getlist('recipients'))
         recipients = [r for r in r_query] # Iterate now to get a list, not a query object
         print "Found recipients %s from request param %s" % (recipients, request.args.getlist('recipients'))
         # Will try to create a new conversation if none already exists
@@ -88,7 +87,6 @@ def conversation_new():
         return redirect(url_for('.conversation_detail', conv_id=conversation.id, modal=modal))
     else:
         abort(400) # No recipients, so who would the conversation be with?
- 
         
 @social.route('/conversations/<conv_id>/', methods=['GET', 'POST'])
 @auth.login_required
@@ -96,8 +94,8 @@ def conversation_detail(conv_id):
     user = auth.get_logged_in_user()
     modal = request.args.has_key('modal')
     
-    conversation = get_object_or_404(Conversation, id=conv_id)
-    messages = Message.select().where(conversation=conv_id).order_by(('pub_date', 'desc'))
+    conversation = get_object_or_404(Conversation, Conversation.id==conv_id)
+    messages = Message.select().where(Message.conversation == conv_id).order_by(Message.pub_date.desc())
     recipients = [r for r in conversation.members()]
     recipients.remove(user) # TODO check for error
     # We are updating a conversation
@@ -115,8 +113,8 @@ def conversation_detail(conv_id):
 @social.route('/groups/')
 def groups():
     user = auth.get_logged_in_user() 
-    my_master_groups = Group.select().join(GroupMember).where(member=user,status=GROUP_MASTER)
-    groups = Group.select().order_by('name')
+    my_master_groups = Group.select().join(GroupMember).where(GroupMember.member == user, GroupMember.status == GROUP_MASTER)
+    groups = Group.select().order_by(Group.name.asc())
     return object_list('social/groups.html', groups, 'groups', my_master_groups=my_master_groups)
 
 @social.route('/groups/new', methods=['GET', 'POST'])
@@ -127,7 +125,7 @@ def group_new():
 @social.route('/groups/<groupslug>/delete', methods=['GET', 'POST'])
 @auth.login_required
 def group_delete(groupslug):
-    group = get_object_or_404(Group, slug=groupslug)
+    group = get_object_or_404(Group, Group.slug == groupslug)
     user = auth.get_logged_in_user()
     masters = list(group.masters())
     #print "if not user=%s or not user in masters=%s or len(list(masters))>1=%s" % (not user, not user in masters, len(masters)>1) 
@@ -145,7 +143,7 @@ def group_delete(groupslug):
 
 @social.route('/groups/<groupslug>/', methods=['GET', 'POST'])
 def group_detail(groupslug):
-    group = get_object_or_404(Group, slug=groupslug)
+    group = get_object_or_404(Group, Group.slug == groupslug)
     return edit_group(request, group)
     
 def edit_group(request, group=None):
@@ -167,8 +165,9 @@ def edit_group(request, group=None):
                 form.populate_obj(group)
                 group.save()
                 if is_newgroup: # Add current user as master for created group
-                    group.addMasters([user]) 
-                edit_group_members(group, request, add=True, master=False)
+                    group.addMasters([user])
+                    user.log("created group %s" % group.slug)
+                edit_group_members(user, group, request, add=True, master=False)
                 flash('Your changes have been saved')
                 return redirect(url_for('.group_detail', groupslug=group.slug))
         return error_response('There were errors with your submission')
@@ -177,13 +176,13 @@ def edit_group(request, group=None):
 @social.route('/groups/<groupslug>/members/<action>', methods=['GET','POST'])
 @auth.login_required
 def group_members_change(groupslug, action):
-    group = get_object_or_404(Group, slug=groupslug)
+    group = get_object_or_404(Group, Group.slug == groupslug)
     user = auth.get_logged_in_user()
     if user not in group.masters():
         return error_response('Logged in user %s is not master of group %s so cannot %s' % (user.username, group.name, action))
     if request.method == 'GET':
-        masters = User.select().where(username__in=request.args.getlist('masters')) if request.args.has_key('masters') else None
-        players = User.select().where(username__in=request.args.getlist('players')) if request.args.has_key('players') else None
+        masters = User.select().where(User.username << request.args.getlist('masters')) if request.args.has_key('masters') else None
+        players = User.select().where(User.username << request.args.getlist('players')) if request.args.has_key('players') else None
         if not masters and not players:
             return error_response('No players or masters given for %s' % action)
         return render_template('includes/change_members.html', \
@@ -192,7 +191,7 @@ def group_members_change(groupslug, action):
     else: # POST
         if not request.form.has_key('masters') and not request.form.has_key('players'):
             return error_response('No players or masters given for %s' % action)
-        changes, err = edit_group_members(group, request, add=(action=="add"), master=False)
+        changes, err = edit_group_members(user, group, request, add=(action=="add"), master=False)
         if err: # There was an error
             return err
         if request.args.get('view') == 'm_group_member_selector' and len(changes)==1: #this option only works if there is one member changed
@@ -208,27 +207,27 @@ def group_members_change(groupslug, action):
         else:
             return error_response('Need to specify which type of view this request comes from')
 
-def edit_group_members(group, r, add, master):
+def edit_group_members(user, group, r, add, master):
     formdata = r.form
     modal = r.args.has_key('modal')
     changes = []
     if formdata.has_key('masters'):
-        masters = list(User.select().where(username__in=formdata.getlist('masters')))
+        masters = list(User.select().where(User.username << formdata.getlist('masters')))
         if not masters:
             return [],error_response('None of the given users %s to %s are recognized' % (formdata.getlist('masters'), 'add' if add else 'remove'))
         masters = group.addMembers(masters, GROUP_MASTER) if add else group.removeMembers(masters)
         if not masters:
             return [],error_response('None of the given users %s could be %s' % (formdata.getlist('masters'), 'added' if add else 'removed'))
-        generate_flash('Added' if add else 'Removed','master',[gm.member.username for gm in masters])
+        user.log(generate_flash('Added' if add else 'Removed','master',[gm.member.username for gm in masters],group.slug))
         changes.extend(masters)
     if formdata.has_key('players'):
-        players = list(User.select().where(username__in=formdata.getlist('players')))
+        players = list(User.select().where(User.username << formdata.getlist('players')))
         if not players:
             return [],error_response('None of the given users %s to %s are recognized' % (formdata.getlist('players'), 'add' if add else 'remove'))
         players = group.addMembers(players, GROUP_PLAYER) if add else group.removeMembers(players)
         if not players:
             return [],error_response('None of the given users %s could be %s' % (formdata.getlist('players'), 'added' if add else 'removed'))
-        generate_flash('Added' if add else 'Removed','player',[gm.member.username for gm in players])
+        user.log(generate_flash('Added' if add else 'Removed','player',[gm.member.username for gm in players],group.slug))
         changes.extend(players)
     return changes, None
 
@@ -253,51 +252,63 @@ def followers():
 def user_list():
     user = auth.get_logged_in_user()
     if user:
-        mastered_groups = Group.select().join(GroupMember).where(member=user,status=GROUP_MASTER)
+        mastered_groups = Group.select().join(GroupMember).where(GroupMember.member == user, GroupMember.status == GROUP_MASTER)
         # For each user in the groups mastered by current user, 
         mastered_members = dict()
-        for gm in GroupMember.select().where(group__in=mastered_groups):
+        for gm in GroupMember.select().where(GroupMember.group << mastered_groups):
             if gm.member.username != user.username: # remove ourselves, no need to check that again!
                 mastered_members['%s_%s' % (gm.member.username,gm.group.id)] = gm
-    users = User.select().order_by('username')
+    users = User.select().order_by(User.username.asc())
     return object_list('social/all_users.html', users, 'user_list', mastered_groups=mastered_groups, mastered_members=mastered_members)
 
-@social.route('/users/<username>/')
+@social.route('/users/<username>/', methods=['GET','POST'])
 def user_detail(username):
-    user = get_object_or_404(User, username=username)
+    user = get_object_or_404(User, User.username == username)
     logged_in = auth.get_logged_in_user()
-    if logged_in:
-        mastered_groups = Group.select().join(GroupMember).where(member=logged_in,status=GROUP_MASTER)
-        # For each user in the groups mastered by current user, 
-        mastered_members = dict()
-        for gm in GroupMember.select().where(group__in=mastered_groups):
-            if gm.member.username != logged_in.username: # remove ourselves, no need to check that again!
-                mastered_members['%s_%s' % (gm.member.username,gm.group.id)] = gm
-    
-    messages = user.message_set.order_by(('pub_date', 'desc'))
-    groups = Group.select().join(GroupMember).where(member=user)
-    return object_list('social/user_detail.html', messages, 'message_list', person=user, groups=groups, 
-        mastered_groups=mastered_groups, mastered_members=mastered_members, modal=request.args.has_key('modal'))
+    edit_allowed = logged_in.id==user.id
+    if request.method == 'GET':
+        if logged_in:
+            mastered_groups = Group.select().join(GroupMember).where(GroupMember.member == logged_in, GroupMember.status == GROUP_MASTER)
+            # For each user in the groups mastered by current user, 
+            mastered_members = dict()
+            for gm in GroupMember.select().where(GroupMember.group << mastered_groups):
+                if gm.member.username != logged_in.username: # remove ourselves, no need to check that again!
+                    mastered_members['%s_%s' % (gm.member.username,gm.group.id)] = gm
+        form = UserForm(obj=user) if edit_allowed else None
+        messages = Message.select().where(Message.user == user).order_by(Message.pub_date.desc())
+        groups = Group.select().join(GroupMember).where(GroupMember.member == user)
+        return object_list('social/user_detail.html', messages, 'message_list', person=user, groups=groups, form=form, 
+            mastered_groups=mastered_groups, mastered_members=mastered_members, modal=request.args.has_key('modal'))
+    elif request.method == 'POST':
+        if not edit_allowed:
+            return error_response('You need to be logged in as %s to edit %s' % (username, username))
+        form = UserForm(request.form, obj=user)
+        if form.validate():
+            form.populate_obj(user)
+            user.save()
+            flash('Your changes have been saved')
+            return redirect(url_for('.user_detail', username=user.username))
+        return error_response('There were errors with your submission')
 
 @social.route('/users/<username>/follow/', methods=['POST'])
 @auth.login_required
 def user_follow(username):
-    user = get_object_or_404(User, username=username)
-    Relationship.get_or_create(
-        from_user=auth.get_logged_in_user(),
-        to_user=user,
-    )
+    user = get_object_or_404(User, User.username == username)
+    logged_in = auth.get_logged_in_user()
+    Relationship.get_or_create(from_user=logged_in, to_user=user)
     flash('You are now following %s' % user.username)
+    logged_in.log('now following %s' % user.username)
+    user.log('now being followed by %s' % logged_in.username)
     # Note point "." before redirect route, it refers to function in this blueprint (e.g social)
     return redirect(url_for('.user_detail', username=user.username, modal=request.args.has_key('modal')))
 
 @social.route('/users/<username>/unfollow/', methods=['POST'])
 @auth.login_required
 def user_unfollow(username):
-    user = get_object_or_404(User, username=username)
-    Relationship.delete().where(
-        from_user=auth.get_logged_in_user(),
-        to_user=user,
-    ).execute()
+    user = get_object_or_404(User, User.username == username)
+    logged_in = auth.get_logged_in_user()
+    Relationship.delete().where(Relationship.from_user==auth.get_logged_in_user(),Relationship.to_user==user).execute()
     flash('You are no longer following %s' % user.username)
+    logged_in.log('now not following %s' % user.username)
+    user.log('was unfollowed by %s' % logged_in.username)
     return redirect(url_for('.user_detail', username=user.username, modal=request.args.has_key('modal')))
